day24/ai-security-platform/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
kafka-python==2.0.2
clickhouse-driver==0.2.6
pymongo==4.6.0
pydantic==2.5.0
pydantic-settings==2.1.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7
websockets==12.0
google-generativeai==0.3.2
grpcio==1.60.0
google-auth==2.25.2
prometheus-client==0.19.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
fakeredis==2.20.1
pytest-mock==3.12.0
coverage==7.3.2
bandit==1.7.5
safety==2.3.5
black==24.8.0
flake8==6.1.0
pre-commit==3.6.0
